Move DNS related parts of traffic manager's State to their own file
Because the state.go file is large and hard to navigate.
